da: braille mismatches, allow for human comments.
After the list of mismatches, there is a closing bracket to end the list 
of mismatches, followed by a set of empty quotes. Human comments can be 
entered here.
Again, the full entry will be skipped if the initial 0 is not changed.

This means that very broken entries can be highlighted by using 
something like
"fixme: all provided alternatives are wrong, but it should be written as 125-24"

The system will not do anything with them, but we can search for fixme 
and see the specifics.



git-svn-id: https://subversion.assembla.com/svn/screenReaderTranslations@11513 080cfd25-d5d9-4c59-82ea-70d02a200d8b